requirements_dev.txt
# pip install -r requirements_dev.txt

# base ----------------------------------------
matplotlib==3.3.4
numpy>=1.18.5
opencv-python>=4.1.2
Pillow~=8.2.0
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
albumentations>=0.5.2
imgaug>=0.4.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas~=1.1.5

# export --------------------------------------
coremltools>=4.1
onnx>=1.6.0
scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
pycocotools>=2.0  # COCO mAP

# server --------------------------------------
grpcio==1.38.0
protobuf==3.17.1
grpcio-tools==1.38.0
requests~=2.25.1